AWSLogsPublisher: Use new class "AWSLogsBuildInfoAction"
