+++ src/main/java/jenkins/plugins/awslogspublisher/AWSLogsPublisher.java
@@ -4,7 +4,6 @@ import hudson.Extension;
 import hudson.Launcher;
 import hudson.model.AbstractBuild;
 import hudson.model.BuildListener;
-import hudson.model.ProminentProjectAction;
 import hudson.tasks.BuildStepMonitor;
 import hudson.tasks.Recorder;
 import org.kohsuke.stapler.DataBoundConstructor;
@@ -51,10 +50,7 @@ public class AWSLogsPublisher extends Recorder {
             return;
         }
 
-        final String url = String.format("https://console.aws.amazon.com/cloudwatch/home?region=%s#logEventViewer:group=%s;stream=%s",
-                config.getAwsRegion(), config.getLogGroupName(), logStreamName);
-
-        addProminentAWSLogsLink(build, url);
+        final String url = addProminentAWSLogsLink(build, config, logStreamName);
 
         try {
             logger.print("[AWS Logs] Build log published at ");
@@ -66,23 +62,12 @@ public class AWSLogsPublisher extends Recorder {
         }
     }
 
-    private static void addProminentAWSLogsLink(AbstractBuild build, final String url) {
-        build.addAction(new ProminentProjectAction() {
-            @Override
-            public String getIconFileName() {
-                return "notepad.png";
-            }
+    private static String addProminentAWSLogsLink(AbstractBuild build, final AWSLogsConfig config, String logStreamName) {
+        final String url = String.format("https://console.aws.amazon.com/cloudwatch/home?region=%s#logEventViewer:group=%s;stream=%s",
+                config.getAwsRegion(), config.getLogGroupName(), logStreamName);
 
-            @Override
-            public String getDisplayName() {
-                return "AWS CloudWatch Logs";
-            }
-
-            @Override
-            public String getUrlName() {
-                return url;
-            }
-        });
+        build.addAction(new AWSLogsBuildInfoAction(config.getLogGroupName(), logStreamName, url));
+        return url;
     }
 
     // Overridden for better type safety.
