+++ src/main/resources/jenkins/plugins/awslogspublisher/AWSLogsBuildInfoAction/summary.jelly
@@ -1,0 +1,7 @@
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout"
+         xmlns:t="/lib/hudson" xmlns:f="/lib/form">
+  <t:summary icon="notepad.png">
+    ${%blurb(it.urlName, it.groupName, it.streamName)}
+  </t:summary>
+</j:jelly>
