+++ src/main/resources/jenkins/plugins/awslogspublisher/AWSLogsBuildInfoAction/summary.properties
@@ -1,0 +1,1 @@
+blurb=AWS CloudWatch Logs: <a href="{0}">{1}:{2}</a>
